requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
ansible
pytest
pytest-cov
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc
loguru
concrete-ml
gradio